finished grid.py component preview
