+++ bread/views/administration.py
@@ -16,7 +16,7 @@ from django.utils.translation import gettext_lazy as _
 from django_celery_results.models import TaskResult
 
 from bread import layout
-from bread.layout.components import grid, tabs, tile
+from bread.layout.components import tabs
 from bread.layout.components.button import Button
 from bread.layout.components.datatable import DataTable, DataTableColumn
 from bread.layout.components.forms import Form, FormField
@@ -139,7 +139,7 @@ def componentpreview(request):
     return tabs.Tabs(
         tabs.Tab(
             _("Layout"),
-            componentpreview_layout(),
+            layout.componentpreview.layout(),
         ),
         tabs.Tab(
             _("Informational"),
@@ -256,242 +256,6 @@ def componentpreview(request):
                             link=Link(href="#", label="link"),
                         ),
                     ),
-                ),
-            ),
-        ),
-    )
-
-
-def componentpreview_layout():
-    def grid_col_preview(content):
-        return grid.Col(
-            content,
-            style=(
-                "align-items: center;"
-                "background: #edf5ff; "
-                "color: #000000; "
-                "display: flex;"
-                "height: 100px; "
-                "justify-content: center; "
-                "outline: 1px solid #a6c8ff; "
-                "overflow: hidden; "
-                "text-align: center;"
-            ),
-        )
-
-    def grid_content_area_preview(content):
-        return hg.DIV(
-            content,
-            style=(
-                "display: flex;"
-                "justify-content: center;"
-                "align-items: center;"
-                "background-color: #9ef0f0;"
-                "height: 100px; "
-                "text-align: center;"
-                "width: 100%; "
-            ),
-        )
-
-    grid_gutter_preview = hg.BaseElement(
-        grid.Row(
-            hg.Iterator(
-                range(1, 5),
-                "colindex",
-                grid_col_preview(hg.F(lambda c: str(c["colindex"]) + " of 4")),
-            )
-        ),
-        grid.Row(
-            hg.Iterator(
-                range(1, 3),
-                "colindex",
-                grid_col_preview(hg.F(lambda c: str(c["colindex"]) + " of 2")),
-            )
-        ),
-        grid.Row(grid_col_preview("1 of 1")),
-    )
-
-    grid_mode_preview = hg.BaseElement(
-        grid.Row(
-            hg.Iterator(
-                range(4),
-                "colindex",
-                grid_col_preview(grid_content_area_preview("area")),
-            )
-        ),
-        grid.Row(
-            hg.Iterator(
-                range(2),
-                "colindex",
-                grid_col_preview(grid_content_area_preview("area")),
-            )
-        ),
-        grid.Row(grid_col_preview(grid_content_area_preview("area"))),
-    )
-
-    def row_mode_preview(gridmode):
-        return grid.Row(
-            hg.Iterator(
-                range(4),
-                "colindex",
-                grid_col_preview(grid_content_area_preview("area")),
-            ),
-            gridmode=gridmode,
-        )
-
-    return hg.BaseElement(
-        hg.H6("bread.layout.components.grid"),
-        hg.H3("Grid", style="margin-bottom: 1.5rem;"),
-        hg.PRE(grid.Grid.__doc__),
-        grid.Row(
-            grid.Col(
-                tile.Tile(
-                    hg.H4("gutter=True (default)"),
-                    grid.Grid(
-                        grid_gutter_preview,
-                        gutter=True,
-                    ),
-                    style="margin-bottom: 1rem;",
-                ),
-            ),
-            grid.Col(
-                tile.Tile(
-                    hg.H4("gutter=False"),
-                    grid.Grid(
-                        grid_gutter_preview,
-                        gutter=False,
-                    ),
-                    style="margin-bottom: 1rem;",
-                ),
-            ),
-        ),
-        grid.Row(
-            grid.Col(
-                tile.Tile(
-                    hg.H4('gridmode="full-width" (default)'),
-                    grid.Grid(
-                        grid_mode_preview,
-                        gridmode="full-width",
-                    ),
-                    style="margin-bottom: 1rem;",
-                ),
-            ),
-            grid.Col(
-                tile.Tile(
-                    hg.H4('gridmode="condensed"'),
-                    grid.Grid(
-                        grid_mode_preview,
-                        gridmode="condensed",
-                    ),
-                    style="margin-bottom: 1rem;",
-                ),
-            ),
-            grid.Col(
-                tile.Tile(
-                    hg.H4('gridmode="narrow"'),
-                    grid.Grid(
-                        grid_mode_preview,
-                        gridmode="narrow",
-                    ),
-                    style="margin-bottom: 1rem;",
-                ),
-            ),
-        ),
-        hg.H6("bread.layout.components.grid"),
-        hg.H3("Row", style="margin-bottom: 1.5rem;"),
-        hg.PRE(grid.Row.__doc__),
-        grid.Row(
-            grid.Col(
-                tile.Tile(
-                    hg.H4("gridmode=None (default)"),
-                    hg.H5('* the same as "full-width"'),
-                    row_mode_preview(None),
-                    style="margin-bottom: 1rem;",
-                ),
-            ),
-            grid.Col(
-                tile.Tile(
-                    hg.H4('gridmode="condensed"'),
-                    row_mode_preview("condensed"),
-                    style="margin-bottom: 1rem;",
-                ),
-            ),
-            grid.Col(
-                tile.Tile(
-                    hg.H4('gridmode="narrow"'),
-                    row_mode_preview("narrow"),
-                    style="margin-bottom: 1rem;",
-                ),
-            ),
-        ),
-        hg.H6("bread.layout.components.grid"),
-        hg.H3("Col", style="margin-bottom: 1.5rem;"),
-        hg.PRE(grid.Col.__doc__),
-        hg.H4(),
-        grid.Row(
-            grid.Col(
-                hg.H4("Flexible width"),
-                tile.Tile(
-                    hg.H4("width=None (default)"),
-                    hg.P(
-                        "Widths are divided equally within the same row regardless of the number of columns inside."
-                    ),
-                    grid.Grid(
-                        hg.Iterator(
-                            range(5, 0, -1),
-                            "rowindex",
-                            grid.Row(
-                                hg.F(
-                                    lambda cout: hg.Iterator(
-                                        range(1, cout["rowindex"] + 1),
-                                        "colindex",
-                                        grid_col_preview(
-                                            hg.F(
-                                                lambda c: "%d of %d"
-                                                % (c["colindex"], c["rowindex"])
-                                            )
-                                        ),
-                                    )
-                                ),
-                            ),
-                        ),
-                    ),
-                    style="margin-bottom: 1rem;",
-                ),
-            ),
-        ),
-        grid.Row(
-            grid.Col(
-                hg.H4("Fixed width"),
-                hg.P(
-                    "You can learn more about possible widths (number of columns) ",
-                    hg.A(
-                        "here",
-                        href="https://www.carbondesignsystem.com/guidelines/2x-grid/overview/#breakpoints",
-                        rel="noreferrer noopener",
-                        target="_blank",
-                    ),
-                    ".",
-                ),
-                tile.Tile(
-                    hg.H4('breakpoint="sm"'),
-                    hg.P(
-                        (
-                            "the maximum total width for 'sm' is 4. "
-                            "Columns that go beyond the total width of their prior columns "
-                            "will fall as a new row instead."
-                        )
-                    ),
-                    grid.Grid(
-                        grid.Row(
-                            hg.Iterator(
-                                range(2),
-                                "colindex",
-                                grid_col_preview("content"),
-                            ),
-                        ),
-                    ),
-                    style="margin-bottom: 1rem;",
                 ),
             ),
         ),
